refactor: extract shared hook utilities to hooks/lib/hook-utils.mjs
Deduplicate readStdin (9 copies), resolveProjectSource + coralProjectDir
(4 copies each), resolveKbRoot (2 copies), and sweepStale (3 copies)
into a single shared ESM module. Net reduction ~165 lines.
