+++ hooks/coral-skill-vars.mjs
@@ -1,8 +1,7 @@
 #!/usr/bin/env node
 
-import { execSync } from 'node:child_process';
-import { homedir } from 'node:os';
-import { basename, join } from 'node:path';
+import { join } from 'node:path';
+import { readStdin, coralProjectDir } from './lib/hook-utils.mjs';
 
 const CORAL_SKILLS = /\/(?:coral:)?(?:plan|preplan|analyze|ralph|bid|discuss|init-project|bugfix|code-simplify)\b/;
 
@@ -40,34 +39,4 @@ try {
   }));
 } catch {
   process.exit(0);
-}
-
-function resolveProjectSource(projectDir) {
-  try {
-    const remote = execSync('git remote get-url origin', {
-      cwd: projectDir,
-      encoding: 'utf-8',
-      stdio: ['pipe', 'pipe', 'pipe'],
-    }).trim().replace(/\.git$/, '');
-    const sshPath = remote.match(/^[^@]+@[^:]+:(.+)$/)?.[1];
-    const rawPath = sshPath ?? remote.replace(/^[^:]+:\/\//, '').replace(/^[^@/]+@/, '').replace(/^[^/]+\/+/, '');
-    const segments = rawPath.split('/').filter(Boolean);
-    if (segments.length >= 2) return `${segments.at(-2)}/${segments.at(-1)}`;
-  } catch {
-    // fall through
-  }
-  return `local/${basename(projectDir)}`;
-}
-
-function coralProjectDir(projectDir) {
-  return join(homedir(), '.coral', 'projects', resolveProjectSource(projectDir).replace(/\//g, '-'));
-}
-
-function readStdin() {
-  return new Promise(resolve => {
-    let data = '';
-    process.stdin.on('data', chunk => { data += chunk; });
-    process.stdin.on('end', () => resolve(data));
-    process.stdin.on('error', () => resolve('{}'));
-  });
 }
