+++ hooks/kb-memo-reminder.mjs
@@ -6,10 +6,10 @@
  * Fail-open: any error exits silently.
  */
 
-import { execSync } from 'node:child_process';
-import { existsSync, mkdirSync, readdirSync, statSync, unlinkSync, writeFileSync } from 'node:fs';
-import { homedir, tmpdir } from 'node:os';
-import { basename, join } from 'node:path';
+import { existsSync, mkdirSync, statSync, writeFileSync } from 'node:fs';
+import { tmpdir } from 'node:os';
+import { join } from 'node:path';
+import { readStdin, coralProjectDir, sweepStale } from './lib/hook-utils.mjs';
 
 const THROTTLE_MIN = 30;
 const FLAG_PREFIX = 'memo-reminded-';
@@ -42,45 +42,4 @@ try {
   }));
 } catch {
   process.exit(0);
-}
-
-function sweepStale(dir, prefix, ttlMs) {
-  try {
-    const now = Date.now();
-    for (const f of readdirSync(dir)) {
-      if (!f.startsWith(prefix)) continue;
-      const p = join(dir, f);
-      if (now - statSync(p).mtimeMs > ttlMs) try { unlinkSync(p); } catch {}
-    }
-  } catch {}
-}
-
-function resolveProjectSource(projectDir) {
-  try {
-    const remote = execSync('git remote get-url origin', {
-      cwd: projectDir,
-      encoding: 'utf-8',
-      stdio: ['pipe', 'pipe', 'pipe'],
-    }).trim().replace(/\.git$/, '');
-    const sshPath = remote.match(/^[^@]+@[^:]+:(.+)$/)?.[1];
-    const rawPath = sshPath ?? remote.replace(/^[^:]+:\/\//, '').replace(/^[^@/]+@/, '').replace(/^[^/]+\/+/, '');
-    const segments = rawPath.split('/').filter(Boolean);
-    if (segments.length >= 2) return `${segments.at(-2)}/${segments.at(-1)}`;
-  } catch {
-    // fall through
-  }
-  return `local/${basename(projectDir)}`;
-}
-
-function coralProjectDir(projectDir) {
-  return join(homedir(), '.coral', 'projects', resolveProjectSource(projectDir).replace(/\//g, '-'));
-}
-
-function readStdin() {
-  return new Promise(resolve => {
-    let data = '';
-    process.stdin.on('data', chunk => { data += chunk; });
-    process.stdin.on('end', () => resolve(data));
-    process.stdin.on('error', () => resolve('{}'));
-  });
 }
