+++ hooks/session-start.mjs
@@ -1,9 +1,8 @@
 #!/usr/bin/env node
 
-import { execSync } from 'node:child_process';
 import { existsSync, mkdirSync, readFileSync, writeFileSync } from 'node:fs';
-import { homedir } from 'node:os';
-import { basename, join } from 'node:path';
+import { join } from 'node:path';
+import { readStdin, resolveProjectSource, coralProjectDir, resolveKbRoot } from './lib/hook-utils.mjs';
 
 try {
   const input = JSON.parse(await readStdin());
@@ -37,33 +36,6 @@ try {
   process.exit(0);
 }
 
-function resolveProjectSource(projectDir) {
-  try {
-    const remote = execSync('git remote get-url origin', {
-      cwd: projectDir,
-      encoding: 'utf-8',
-      stdio: ['pipe', 'pipe', 'pipe'],
-    }).trim().replace(/\.git$/, '');
-    const sshPath = remote.match(/^[^@]+@[^:]+:(.+)$/)?.[1];
-    const rawPath = sshPath ?? remote.replace(/^[^:]+:\/\//, '').replace(/^[^@/]+@/, '').replace(/^[^/]+\/+/, '');
-    const segments = rawPath.split('/').filter(Boolean);
-    if (segments.length >= 2) return `${segments.at(-2)}/${segments.at(-1)}`;
-  } catch {
-    // fall through
-  }
-  return `local/${basename(projectDir)}`;
-}
-
-function coralProjectDir(projectDir) {
-  return join(homedir(), '.coral', 'projects', resolveProjectSource(projectDir).replace(/\//g, '-'));
-}
-
-function resolveKbRoot() {
-  const custom = process.env.CORAL_KB_PATH;
-  if (custom) return custom.startsWith('~') ? join(homedir(), custom.slice(1)) : custom;
-  return join(homedir(), '.coral', 'kb');
-}
-
 function ensureCliPermission(projectDir) {
   const rule = 'Bash(node *coral-cli*)';
   const dir = join(projectDir, '.claude');
@@ -80,13 +52,4 @@ function ensureCliPermission(projectDir) {
   } catch {
     // fail-open
   }
-}
-
-function readStdin() {
-  return new Promise(resolve => {
-    let data = '';
-    process.stdin.on('data', chunk => { data += chunk; });
-    process.stdin.on('end', () => resolve(data));
-    process.stdin.on('error', () => resolve('{}'));
-  });
 }
